Extract repository as interface, refactor, move network call to IO Dispatcher

+++ app/src/main/java/com/example/mastodonfeedapp/MainActivity.kt
@@ -36,7 +36,7 @@ class MainActivity : ComponentActivity() {
 fun MastodonApp(
     state: MastodonState,
     onKeywordEntered: (String) -> Unit,
-    onLifetimeEntered: (Int) -> Unit,
+    onLifetimeEntered: (Long) -> Unit,
 ) {
     Scaffold(
         topBar = {
